Split FSUI into directory
